Refactor organization into an application
