--- src/api-engine/api/routes/file/__init__.py
@@ -1,1 +1,0 @@
-#
