--- src/api-engine/api/routes/organization/__init__.py
@@ -1,3 +1,0 @@
-#
-# SPDX-License-Identifier: Apache-2.0
-#
