+++ src/api-engine/api/utils/common.py
@@ -89,18 +89,6 @@ def to_form_paras(self):
     return custom_paras
 
 
-def any_of(*perm_classes):
-    """Returns permission class that allows access for
-    one of permission classes provided in perm_classes"""
-
-    class Or(BasePermission):
-        def has_permission(*args):
-            allowed = [p.has_permission(*args) for p in perm_classes]
-            return reduce(lambda x, y: x or y, allowed)
-
-    return Or
-
-
 def hash_file(file, block_size=65536):
     hash_func = hashlib.md5()
     for buf in iter(partial(file.read, block_size), b""):
